resolves #44: References to HSM and challenge-response protocol
Added references to the SHE+ standard and a link to the AUTOSAR spec. Also added a reference to the three phase challenge response protocol specified in ISO 9798-4, for when a remote device like a touch screen sends a mode change command to the security gateway.

There are other ways to achieve this requirement, of course, but I think this is the most appropriate. The messaging does need protection against replay attacks, which can be done by the results of the cited authentication protocol.